requirements.txt
alembic==1.8.1
aniso8601==9.0.1
bcrypt==3.2.2
blinker==1.5
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Mail==0.9.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-jwt==3.3.3
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
wrapt==1.14.1
